add deprecation warnings
