+++ examples/02_preprocessing/plot_fmri_to_template.py
@@ -26,11 +26,11 @@ output_dir = os.path.abspath('zurich_template')
 
 ##############################################################################
 # We encapsulate them through an object called `FMRISession`
-from sammba.registration import fmri
+from sammba.registration import FuncSession
 
-animal_session = fmri.FMRISession(anat=anat_filename, func=func_filename,
-                                  animal_id='1366', brain_volume=400, t_r=1.,
-                                  output_dir=output_dir)
+animal_session = FuncSession(anat=anat_filename, func=func_filename,
+                             animal_id='1366', brain_volume=400, t_r=1.,
+                             output_dir=output_dir)
 
 ##############################################################################
 # Choose the template
