+++ sammba/registration/__init__.py
@@ -2,7 +2,10 @@ from .utils import fix_obliquity, create_pipeline_graph
 from .func import fmri_sessions_to_template, coregister_fmri_session
 from .struct import anats_to_common, anats_to_template
 from .fmri_session import FMRISession
+from .func import FuncSession
+from .cest import CESTSession
 
 __all__ = ['fix_obliquity', 'create_pipeline_graph',
            'fmri_sessions_to_template', 'anats_to_common', 'FMRISession',
-           'anats_to_template', 'coregister_fmri_session']
+           'anats_to_template', 'coregister_fmri_session',
+           'FuncSession', 'CESTSession']
