+++ sammba/registration/func.py
@@ -1,15 +1,434 @@
 import os
 import numpy as np
 import nibabel
+from sklearn.utils import deprecated
+from nilearn._utils.compat import _basestring
 from sammba.externals.nipype.caching import Memory
 from sammba.externals.nipype.interfaces import afni
 from sammba.externals.nipype.utils.filemanip import fname_presuffix
 from sammba.interfaces import segmentation
 from .utils import fix_obliquity, copy_geometry
-from .fmri_session import FMRISession
 from .struct import anats_to_template
+from .multimodal import (extract_brain, _rigid_body_register, _warp,
+                         _per_slice_qwarp, _transform_to_template)
 
 
+def _realign(func_filename, write_dir, caching=False,
+             terminal_output='allatonce', environ={}):
+    if caching:
+        memory = Memory(write_dir)
+        clip_level = memory.cache(afni.ClipLevel)
+        calc = memory.cache(afni.Calc)
+        volreg = memory.cache(afni.Volreg)
+        allineate = memory.cache(afni.Allineate)
+        tstat = memory.cache(afni.TStat)
+        for step in [volreg, allineate, tstat, calc]:
+            step.interface().set_default_terminal_output(terminal_output)
+    else:
+        clip_level = afni.ClipLevel().run
+        calc = afni.Calc(terminal_output=terminal_output).run
+        volreg = afni.Volreg(terminal_output=terminal_output).run
+        allineate = afni.Allineate(terminal_output=terminal_output).run
+        tstat = afni.TStat(terminal_output=terminal_output).run
+
+    out_clip_level = clip_level(in_file=func_filename)
+    out_calc_threshold = calc(
+        in_file_a=func_filename,
+        expr='ispositive(a-{0}) * a'.format(out_clip_level.outputs.clip_val),
+        outputtype='NIFTI_GZ')
+    thresholded_filename = out_calc_threshold.outputs.out_file
+
+    out_volreg = volreg(  # XXX dfile not saved
+        in_file=thresholded_filename,
+        outputtype='NIFTI_GZ',
+        environ=environ,
+        oned_file=fname_presuffix(thresholded_filename,
+                                  suffix='Vr.1Dfile.1D', use_ext=False),
+        oned_matrix_save=fname_presuffix(thresholded_filename,
+                                         suffix='Vr.aff12.1D',
+                                         use_ext=False))
+
+    # Apply the registration to the whole head
+    out_allineate = allineate(in_file=func_filename,
+                              master=func_filename,
+                              in_matrix=out_volreg.outputs.oned_matrix_save,
+                              out_file=fname_presuffix(func_filename,
+                                                       suffix='Av'),
+                              environ=environ)
+
+    # 3dAllineate removes the obliquity. This is not a good way to readd it as
+    # removes motion correction info in the header if it were an AFNI file...as
+    # it happens it's NIfTI which does not store that so irrelevant!
+    allineated_filename = copy_geometry(
+        filename_to_copy=out_volreg.outputs.out_file,
+        filename_to_change=out_allineate.outputs.out_file)
+
+    # Create a (hopefully) nice mean image for use in the registration
+    out_tstat = tstat(in_file=allineated_filename, args='-mean',
+                      outputtype='NIFTI_GZ', environ=environ)
+
+    # Remove intermediate outputs
+    if not caching:
+        for output_file in [thresholded_filename,
+                            out_volreg.outputs.oned_matrix_save,
+                            out_volreg.outputs.out_file,
+                            out_volreg.outputs.md1d_file]:
+            os.remove(output_file)
+    return (allineated_filename, out_tstat.outputs.out_file,
+            out_volreg.outputs.oned_file)
+
+
+class FuncSession(object):
+    """
+    Encapsulation for fMRI data, relative to preprocessing.
+
+    Parameters
+    ----------
+    func : str
+        Path to the functional 4D image
+
+    anat : str
+        Path to anatomical image
+
+    animal_id : str
+        Animal id
+
+    brain_volume : int
+        Volume of the brain used for brain extraction.
+        Typically 400 for mouse and 1800 for rat.
+
+    t_r : float, optional
+        Repetition time for the EPI, in seconds.
+
+    output_dir : str, optional
+        Path to the output directory. If not specified, current directory is
+        used. Final and intermediate images are stored in the subdirectory
+        `animal_id` of the given `output_dir`.
+    """
+
+    def __init__(self, func=None, anat=None, brain_volume=None,
+                 animal_id=None, t_r=None, output_dir=None):
+        self.func = func
+        self.anat = anat
+        self.brain_volume = brain_volume
+        self.animal_id = animal_id
+        self.t_r = t_r
+        self.output_dir = output_dir
+
+    def _set_items(self, **kwargs):
+        for k, v in kwargs.items():
+            setattr(self, k, v)
+
+    def _check_inputs(self):
+        if not os.path.isfile(self.func):
+            raise IOError('func must be an existing image file,'
+                          'you gave {0}'.format(self.func))
+
+        if not os.path.isfile(self.anat):
+            raise IOError('anat must be an existing image file,'
+                          'you gave {0}'.format(self.anat))
+
+        if self.brain_volume is None:
+            raise ValueError('you must provide the expected brain volume.')
+
+        if not isinstance(self.animal_id, _basestring):
+            raise ValueError('animal_id must be a string, you provided '
+                             '{0}'.format(self.animal_id))
+
+    def _set_output_dir(self):
+        if self.output_dir is None:
+            self.output_dir = os.getcwd()
+
+        if not os.path.isdir(self.output_dir):
+            os.makedirs(self.output_dir)
+
+    def coregister(self, use_rats_tool=True,
+                   slice_timing=True,
+                   prior_rigid_body_registration=False,
+                   caching=False, voxel_size_x=.1, voxel_size_y=.1,
+                   verbose=True, **environ_kwargs):
+        """
+        Coregistration of the subject's functional and anatomical images.
+        The functional volume is aligned to the anatomical, first with a
+        rigid body registration and then on a per-slice basis (only a fine
+        correction, this is mostly for correction of EPI distortion).
+
+        Parameters
+        ----------
+        use_rats_tool : bool, optional
+            If True, brain mask is computed using RATS Mathematical Morphology.
+            Otherwise, a histogram-based brain segmentation is used.
+
+        prior_rigid_body_registration : bool, optional
+            If True, a rigid-body registration of the anat to the func is
+            performed prior to the warp. Useful if the images headers have
+            missing/wrong information.
+
+        voxel_size_x : float, optional
+            Resampling resolution for the x-axis, in mm.
+
+        voxel_size_y : float, optional
+            Resampling resolution for the y-axis, in mm.
+
+        caching : bool, optional
+            Wether or not to use caching.
+
+        verbose : bool, optional
+            If True, all steps are verbose. Note that caching implies some
+            verbosity in any case.
+
+        environ_kwargs : extra arguments keywords
+            Extra arguments keywords, passed to interfaces environ variable.
+
+        Returns
+        -------
+        The following attributes are added
+            - `coreg_func_` : str
+                              Path to paths to the coregistered functional
+                              image.
+            - `coreg_anat_` : str
+                              Path to paths to the coregistered functional
+                              image.
+            - `coreg_transform_` : str
+                                   Path to the transform from anat to func.
+        Notes
+        -----
+        If `use_rats_tool` is turned on, RATS tool is used for brain extraction
+        and has to be cited. For more information, see
+        `RATS <http://www.iibi.uiowa.edu/content/rats-overview/>`_
+        """
+        func_filename = self.func
+        anat_filename = self.anat
+
+        environ = {'AFNI_DECONFLICT': 'OVERWRITE'}
+        for (key, value) in environ_kwargs.items():
+            environ[key] = value
+
+        if verbose:
+            terminal_output = 'allatonce'
+        else:
+            terminal_output = 'none'
+
+        if caching:
+            memory = Memory(self.output_dir)
+            tshift = memory.cache(afni.TShift)
+            unifize = memory.cache(afni.Unifize)
+            catmatvec = memory.cache(afni.CatMatvec)
+            for step in [tshift, unifize]:
+                step.interface().set_default_terminal_output(terminal_output)
+            overwrite = False
+        else:
+            tshift = afni.TShift(terminal_output=terminal_output).run
+            unifize = afni.Unifize(terminal_output=terminal_output).run
+            catmatvec = afni.CatMatvec().run
+            overwrite = True
+
+        self._check_inputs()
+        self._set_output_dir()
+        images_dir = os.path.join(os.path.abspath(self.output_dir),
+                                  self.animal_id)
+        current_dir = os.getcwd()
+        os.chdir(images_dir)
+        output_files = []
+
+        #######################################
+        # Correct functional for slice timing #
+        #######################################
+        if slice_timing:
+            if self.t_r is None:
+                raise ValueError('`t_r` is needed for slice timing correction')
+
+            out_tshift = tshift(in_file=func_filename,
+                                outputtype='NIFTI_GZ',
+                                tpattern='altplus',
+                                tr=str(self.t_r),
+                                environ=environ)
+            func_filename = out_tshift.outputs.out_file
+            output_files.append(func_filename)
+
+        ################################################
+        # Register functional volumes to the first one #
+        ################################################
+        allineated_filename, mean_aligned_filename, _ = \
+            _realign(func_filename, self.output_dir, caching=caching,
+                     terminal_output=terminal_output, environ=environ)
+
+        # Update outputs
+        output_files.extend([allineated_filename, mean_aligned_filename])
+        ###########################################
+        # Corret anat and func for intensity bias #
+        ###########################################
+        # Correct the functional average for intensities bias
+        out_bias_correct = unifize(in_file=mean_aligned_filename,
+                                   outputtype='NIFTI_GZ', environ=environ)
+        unbiased_mean_func_filename = out_bias_correct.outputs.out_file
+
+        # Bias correct the antomical image
+        out_unifize = unifize(in_file=anat_filename, outputtype='NIFTI_GZ',
+                              environ=environ)
+        unbiased_anat_filename = out_unifize.outputs.out_file
+
+        # Update outputs
+        output_files.extend([unbiased_mean_func_filename,
+                             unbiased_anat_filename])
+
+        #############################################
+        # Rigid-body registration anat -> mean func #
+        #############################################
+        if prior_rigid_body_registration:
+            anat_brain_filename = extract_brain(
+                unbiased_anat_filename, self.output_dir, self.brain_volume,
+                caching=False, use_rats_tool=use_rats_tool,
+                terminal_output=terminal_output, environ=environ)
+            func_brain_filename = extract_brain(
+                unbiased_mean_func_filename, self.output_dir,
+                self.brain_volume, caching=False, use_rats_tool=use_rats_tool,
+                terminal_output=terminal_output, environ=environ)
+            allineated_anat_filename, rigid_transform_file = \
+                _rigid_body_register(unbiased_anat_filename,
+                                     anat_brain_filename,
+                                     unbiased_mean_func_filename,
+                                     func_brain_filename, self.output_dir,
+                                     caching=False,
+                                     terminal_output=terminal_output,
+                                     environ=environ)
+            output_files.extend([rigid_transform_file,
+                                 allineated_anat_filename])
+        else:
+            allineated_anat_filename = unbiased_anat_filename
+
+        ############################################
+        # Nonlinear registration anat -> mean func #
+        ############################################
+        registered_anat_oblique_filename, mat_filename, warp_output_files =\
+            _warp(allineated_anat_filename, unbiased_mean_func_filename,
+                  self.output_dir, caching=caching,
+                  terminal_output=terminal_output, overwrite=overwrite,
+                  environ=environ)
+        # Concatenate all the anat to func tranforms
+        output_files.extend(warp_output_files)
+        transform_filename = fname_presuffix(registered_anat_oblique_filename,
+                                             suffix='_anat_to_func.aff12.1D',
+                                             use_ext=False)
+        _ = catmatvec(in_file=[(mat_filename, 'ONELINE')],
+                      oneline=True,
+                      out_file=transform_filename)
+
+        ##################################################
+        # Per-slice non-linear registration func -> anat #
+        ##################################################
+        warped_mean_func_filename, warp_filenames, warped_func_filename =\
+            _per_slice_qwarp(unbiased_mean_func_filename,
+                             registered_anat_oblique_filename,
+                             self.output_dir, voxel_size_x, voxel_size_y,
+                             apply_to_file=allineated_filename,
+                             overwrite=overwrite,
+                             caching=caching, terminal_output=terminal_output,
+                             environ=environ)
+        # Update the outputs
+        output_files.append(warped_mean_func_filename)
+        if not caching:
+            for out_file in output_files:
+                os.remove(out_file)
+
+        # Update the fmri data
+        setattr(self, "coreg_func_", warped_func_filename)
+        setattr(self, "coreg_anat_", registered_anat_oblique_filename)
+        setattr(self, "coreg_transform_", transform_filename)
+        os.chdir(current_dir)
+
+    def register_to_template(self, head_template_filename,
+                             brain_template_filename=None,
+                             dilated_head_mask_filename=None,
+                             prior_rigid_body_registration=False,
+                             slice_timing=True,
+                             func_voxel_size=None,
+                             maxlev=None,
+                             caching=False, verbose=True):
+        """ Registration of subject's functional and anatomical images to
+        a given template.
+
+        Parameters
+        ----------
+        head_template_filename : str
+            Template to register the functional to.
+
+        brain_template_filename : str, optional
+            Path to a brain template, passed to
+            sammba.registration.anats_to_template
+
+        dilated_head_mask_filename : str, optional
+            Path to a dilated head mask, passed to
+            sammba.registration.anats_to_template
+
+        func_voxel_size : 3-tuple of floats, optional
+            Voxel size of the registered functional, in mm.
+
+        maxlev : int or None, optional
+            Maximal level for the warp when registering anat to template.
+            Passed to
+            sammba.registration.anats_to_template
+
+        caching : bool, optional
+            Wether or not to use caching.
+
+        verbose : bool, optional
+            If True, all steps are verbose. Note that caching implies some
+            verbosity in any case.
+
+        Returns
+        -------
+        The following attributes are added/updated
+            - `template_` : str
+                           Path to the given registration template.
+            - `registered_func_` : str
+                                   Path to the funct registered to template.
+            - `registered_anat_` : str
+                                   Path to the anat registered to template.
+
+        See also
+        --------
+        sammba.registration.anats_to_template
+        """
+        self._check_inputs()
+        images_dir = os.path.join(os.path.abspath(self.output_dir),
+                                  self.animal_id)
+
+        if not hasattr(self, 'coreg_func_') or not hasattr(self,
+                                                           'coreg_transform_'):
+            raise ValueError('Anatomical and functional have not been '
+                             'coregistered. Please use `coreg` function first')
+
+        # XXX do a function for creating new attributes ?
+        setattr(self, "template_", head_template_filename)
+        anats_registration = anats_to_template(
+            [self.anat],
+            head_template_filename,
+            images_dir,
+            self.brain_volume,
+            brain_template_filename=brain_template_filename,
+            dilated_head_mask_filename=dilated_head_mask_filename,
+            maxlev=maxlev,
+            caching=caching, verbose=verbose)
+        setattr(self, "registered_anat_", anats_registration.registered[0])
+
+        normalized_func_filename = _transform_to_template(
+            self.coreg_func_,
+            head_template_filename,
+            images_dir,
+            [self.coreg_transform_, anats_registration.pre_transforms[0],
+             anats_registration.transforms[0]],
+            voxel_size=func_voxel_size, caching=caching, verbose=verbose)
+
+        setattr(self, "registered_func_", normalized_func_filename)
+
+
+from .fmri_session import FMRISession
+
+
+@deprecated("Function 'coregister_fmri_session' has been replaced by "
+            "method 'coregister' of class 'FuncSession' and will be removed "
+            "in future release. ")
 def coregister_fmri_session(session_data, t_r, write_dir, brain_volume,
                             use_rats_tool=True,
                             slice_timing=True,
@@ -505,6 +924,9 @@ def coregister_fmri_session(session_data, t_r, write_dir, brain_volume,
                 os.remove(out_file)
 
 
+@deprecated("Function 'fmri_sessions_to_template' has been replaced by "
+            "function '_transform_to_template' and will be "
+            "removed in future release. ")
 def _func_to_template(func_coreg_filename, template_filename, write_dir,
                       func_to_anat_oned_filename,
                       anat_to_template_oned_filename,
@@ -582,6 +1004,9 @@ def _func_to_template(func_coreg_filename, template_filename, write_dir,
     return normalized_filename
 
 
+@deprecated("Function 'fmri_sessions_to_template' has been replaced by "
+            "method 'register_to_template' of class 'FuncSession' and will be "
+            "removed in future release. ")
 def fmri_sessions_to_template(sessions, t_r, head_template_filename,
                               write_dir,
                               brain_volume, use_rats_tool=True,
